requirements.txt
requests
allure-pytest
pytest
PyHamcrest